Alex™ File Deleter action #1
From providing the deletion for the chip to delivering the deletion services that allow organizations to securely manage the deletion of products throughout their life cycle, Alex delivers a complete Deletion of Things (DoT) solution for our partners and customers. It’s rooted in our deep understanding of the future of destruction and recycling.